requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.7
noise==1.2.2
numpy==2.3.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pygame-ce==2.5.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.1
python-dateutil==2.9.0.post0
six==1.17.0
